% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.meas.R
\name{auroc}
\alias{auroc}
\alias{auroc.single.class}
\alias{auroc.single.over.classes}
\title{AUROC measures}
\usage{
auroc.single.class(labels, scores, folds = NULL, seed = NULL)

auroc.single.over.classes(target, predicted, folds = NULL, seed = NULL)
}
\arguments{
\item{labels}{vector of the true labels (0 negative, 1 positive examples).}

\item{scores}{a numeric vector of the values of the predicted labels (scores).}

\item{folds}{number of folds on which computing the AUROC. If \code{folds=NULL} (\code{def.}), the AUROC is computed one-shot,
otherwise the AUROC is computed averaged across folds.}

\item{seed}{initialization seed for the random generator to create folds. Set \code{seed} only if \code{folds}\eqn{\neq}\code{NULL}.
If \code{seed=NULL} and \code{folds}\eqn{\neq}\code{NULL}, the AUROC averaged across folds is computed without seed initialization.}

\item{target}{annotation matrix: rows correspond to examples and columns to classes.
\eqn{target[i,j]=1} if example \eqn{i} belongs to class \eqn{j}, \eqn{target[i,j]=0} otherwise.}

\item{predicted}{a numeric matrix with predicted values (scores): rows correspond to examples and columns to classes.}
}
\value{
\code{auroc.single.class} returns a numeric value corresponding to the AUROC for the considered class;
\code{auprc.single.over.classes} returns a list with two elements:
\enumerate{
 \item average: the average AUROC across classes;
 \item per.class: a named vector with AUROC for each class. Names correspond to classes.
}
}
\description{
Compute the Area under the ROC Curve (AUROC) through \pkg{precrec} package.
}
\details{
The AUROC (for a single class or for a set of classes) is computed either one-shot or averaged across stratified folds.

\code{auroc.single.class} computes the AUROC just for a given class.

\code{auroc.single.over.classes} computes the AUROC for a set of classes, including their average values across all the classes.

For all those classes having zero annotations, the AUROC is set to 0.5. These classes are included in the computing of the AUROC
averaged across classes, both when the AUROC is computed one-shot or averaged across stratified folds.

The AUROC is set to 0.5 to all those classes having zero annotations.
Names of rows and columns of \code{labels} and \code{predicted} must be provided in the same order, otherwise a stop message is returned.
}
\examples{
data(labels);
data(scores);
data(graph);
root <- root.node(g);
L <- L[,-which(colnames(L)==root)];
S <- S[,-which(colnames(S)==root)];
auc.single.class <- auroc.single.class(L[,3], S[,3], folds=5, seed=23);
auc.over.classes <- auroc.single.over.classes(L, S, folds=5, seed=23);
}
