% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.meas.R
\name{AUROC.single.class}
\alias{AUROC.single.class}
\title{AUROC single class}
\usage{
AUROC.single.class(target, pred)
}
\arguments{
\item{target}{vector of the true labels (0 negative, 1 positive examples)}

\item{pred}{numeric vector of the values of the predicted labels (scores)}
}
\value{
a numeric value corresponding to the AUROC for the considered class
}
\description{
High-level function to compute the Area under the ROC Curve (AUPRC) just for a single class through \pkg{precrec} package
}
\examples{
data(labels);
data(scores);
data(graph);
root <- root.node(g);
L <- L[,-which(colnames(L)==root)];
S <- S[,-which(colnames(S)==root)];
AUC <- AUROC.single.class(L[,3],S[,3]);
}
\seealso{
\code{\link{AUROC.single.over.classes}}
}
