% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.utility.R
\name{build.submatrix}
\alias{build.submatrix}
\title{Build submatrix}
\usage{
build.submatrix(ann, n)
}
\arguments{
\item{ann}{the annotation matrix (0/1). Rows are examples and columns are functional terms.}

\item{n}{an integer number representing the number of annotations to be pruned.}
}
\value{
An annotation matrix having only those terms with more than n annotations.
}
\description{
Terms having less than n annotations are pruned. Terms having exactly n annotations are discarded as well.
}
\examples{
data(labels);
subm <- build.submatrix(L,5);
}
