% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.utility.R
\name{compute.flipped.graph}
\alias{compute.flipped.graph}
\title{Flip graph}
\usage{
compute.flipped.graph(g)
}
\arguments{
\item{g}{a \code{graphNEL} directed graph}
}
\value{
A graph (as an object of class \code{graphNEL}) with edges in the opposite direction w.r.t. \code{g}.
}
\description{
Compute a directed graph with edges in the opposite direction.
}
\examples{
data(graph);
g.flipped <- compute.flipped.graph(g);
}
