% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.utility.R
\name{distances.from.leaves}
\alias{distances.from.leaves}
\title{Distances from leaves}
\usage{
distances.from.leaves(g)
}
\arguments{
\item{g}{a graph of class \code{graphNEL}. It represents the hierarchy of the classes.}
}
\value{
A named vector. The names are the names of the nodes of the graph \code{g}, and their values represent the distance from the leaves.
A value equal to 0 is assigned to the leaves, 1 to nodes with distance 1 from a leaf and so on.
}
\description{
Compute the minimum distance of each node from one of the leaves of the graph.
}
\examples{
data(graph);
dist.leaves <- distances.from.leaves(g);
}
