% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEMDAG.1.0.0.R
\name{do.subgraph}
\alias{do.subgraph}
\title{Build subgraph}
\usage{
do.subgraph(nd, g, edgemode = "directed")
}
\arguments{
\item{nd}{a vector with the nodes for which the subgraph must be built}

\item{g}{a graph of class \code{graphNEL}. It represents the hierarchy of the classes.}

\item{edgemode}{can be "directed" or "undirected"}
}
\value{
a subgraph with only the supplied nodes
}
\description{
This function returns a subgraph with only the supplied nodes and any edges between them
}
\examples{
data(graph);
anc <- build.ancestors(g);
nd <- anc[["HP:0001371"]];
subg <- do.subgraph(nd, g, edgemode="directed");
}
