% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.utility.R
\name{do.submatrix}
\alias{do.submatrix}
\title{Build submatrix}
\usage{
do.submatrix(ann, n)
}
\arguments{
\item{ann}{the annotations matrix (0/1). Rows are examples and columns are classes}

\item{n}{integer number of annotations to be pruned}
}
\value{
Matrix of annotations having only those terms with more than n annotations
}
\description{
Terms having less than n annotations are pruned. Terms having exactly n annotations are discarded as well.
}
\examples{
data(labels);
subm <- do.submatrix(L,5);
}
