% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.meas.R
\name{find.best.f}
\alias{find.best.f}
\title{Best hierarchical F-score}
\usage{
find.best.f(
  target,
  predicted,
  n.round = 3,
  verbose = TRUE,
  b.per.example = FALSE
)
}
\arguments{
\item{target}{matrix with the target multilabel: rows correspond to examples and columns to classes.
\eqn{target[i,j]=1} if example \eqn{i} belongs to class \eqn{j}, \eqn{target[i,j]=0} otherwise.}

\item{predicted}{a numeric matrix with continuous predicted values (scores): rows correspond to examples and columns to classes.}

\item{n.round}{number of rounding digits to be applied to predicted (\code{default=3}).}

\item{verbose}{a boolean value. If \code{TRUE} (def.) the number of iterations are printed on stdout.}

\item{b.per.example}{a boolean value.
\itemize{
 \item \code{TRUE}: results are returned for each example;
 \item \code{FALSE}: only the average results are returned;
}}
}
\value{
Two different outputs respect to the input parameter \code{b.per.example}:
\itemize{
 \item \code{b.per.example==FALSE}: a list with a single element average. A named vector with 7 elements relative to the best result in terms
  of the F.measure: Precision (P), Recall (R), Specificity (S), F.measure (F), av.F.measure (av.F), Accuracy (A) and the best selected Threshold (T).
  F is the F-measure computed as the harmonic mean between the average precision and recall; av.F is the F-measure computed as the average across
  examples and T is the best selected threshold;
 \item \code{b.per.example==FALSE}: a list with two elements:
  \enumerate{
   \item average: a named vector with with 7 elements relative to the best result in terms of the F.measure: Precision (P), Recall (R),
    Specificity (S), F.measure (F), av.F.measure (av.F), Accuracy (A) and the best selected Threshold (T);
   \item per.example: a named matrix with the Precision (P), Recall (R), Specificity (S), Accuracy (A), F-measure (F), av.F-measure (av.F)
    and the best selected Threshold (T) for each example. Row names correspond to examples, column names correspond respectively
    to Precision (P), Recall (R), Specificity (S), Accuracy (A), F-measure (F), av.F-measure (av.F) and the best selected Threshold (T);
  }
}
}
\description{
Select the best hierarchical F-score by choosing an appropriate threshold in the scores.
}
\details{
All the examples having no positive annotations are discarded. The predicted scores matrix (\code{predicted}) is rounded
according to parameter \code{n.round} and all the values of \code{predicted} are divided by \code{max(predicted)}.
Then all the thresholds corresponding to all the different values included in \code{predicted} are attempted, and the threshold
leading to the maximum F-measure is selected.

Names of rows and columns of \code{target} and \code{predicted} matrix must be provided in the same order, otherwise a stop message is returned.
}
\examples{
data(graph);
data(labels);
data(scores);
root <- root.node(g);
L <- L[,-which(colnames(L)==root)];
S <- S[,-which(colnames(S)==root)];
fscore <- find.best.f(L, S, n.round=3, verbose=TRUE, b.per.example=TRUE);
}
