% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.utility.R
\name{find.leaves}
\alias{find.leaves}
\title{Leaves}
\usage{
find.leaves(g)
}
\arguments{
\item{g}{a graph of class \code{graphNEL}. It represents the hierarchy of the classes.}
}
\value{
A vector with the names of the leaves of g.
}
\description{
Find leaves of a directed graph.
}
\examples{
data(graph);
leaves <- find.leaves(g);
}
