% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEMDAG-defunct.R
\name{HEMDAG-defunct}
\alias{HEMDAG-defunct}
\alias{descens.threshold}
\alias{descens.threshold.free}
\alias{descens.weighted.threshold.free}
\alias{descens.weighted.threshold}
\alias{descens.tau}
\alias{tpr.threshold}
\alias{tpr.threshold.free}
\alias{tpr.weighted.threshold.free}
\alias{tpr.weighted.threshold}
\title{Defunct functions in package \pkg{HEMDAG}.}
\usage{
descens.threshold(S, g, root = "00", t = 0.5)

descens.threshold.free(S, g, root = "00")

descens.weighted.threshold.free(S, g, root = "00", w = 0.5)

descens.weighted.threshold(S, g, root = "00", t = 0.5, w = 0.5)

descens.tau(S, g, root = "00", t = 0.5)

tpr.threshold(S, g, root = "00", t = 0.5)

tpr.threshold.free(S, g, root = "00")

tpr.weighted.threshold.free(S, g, root = "00", w = 0.5)

tpr.weighted.threshold(S, g, root = "00", t = 0.5, w = 0.5)
}
\arguments{
\item{S}{a named flat scores matrix with examples on rows and classes on columns}

\item{g}{a graph of class \code{graphNEL}. It represents the hierarchy of the classes}

\item{root}{name of the class that it is on the top-level of the hierarchy (\code{def. root="00"})}

\item{t}{threshold for the choice of the positive descendants (\code{def. t=0.5}); whereas in the \code{descens.tau} variant 
the parameter \code{t} balances the contribution between the positives children of a node \eqn{i} and that of its
positives descendants excluding its positives children}

\item{w}{weight to balance between the contribution of the node \eqn{i} and that of its positive descendants}
}
\value{
a named matrix with the scores of the classes corrected according to the chosen hierarchical variant.
}
\description{
The functions listed below are defunct. The alternative function is supplied.
Help page for defunct functions is available by typing \code{help("HEMDAG-defunct")}
}
\section{\code{descens.threshold}}{
 For \code{descens.threshold}, use \code{\link{TPR.DAG}}.
}

\section{\code{descens.threshold.free}}{
 For \code{descens.threshold.free}, use \code{\link{TPR.DAG}}.
}

\section{\code{descens.weighted.threshold.free}}{
 For \code{descens.weighted.threshold.free}, use \code{\link{TPR.DAG}}.
}

\section{\code{descens.weighted.threshold}}{
 For \code{descens.weighted.threshold}, use \code{\link{TPR.DAG}}.
}

\section{\code{descens.tau}}{
 For \code{descens.tau}, use \code{\link{TPR.DAG}}.
}

\section{\code{tpr.threshold}}{
 For \code{tpr.threshold}, use \code{\link{TPR.DAG}}.
}

\section{\code{tpr.threshold.free}}{
 For \code{tpr.threshold.free}, use \code{\link{TPR.DAG}}.
}

\section{\code{tpr.weighted.threshold.free}}{
 For \code{tpr.weighted.threshold.free}, use \code{\link{TPR.DAG}}.
}

\section{\code{tpr.weighted.threshold}}{
 For \code{tpr.weighted.threshold}, use \code{\link{TPR.DAG}}.
}

