% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.fun.R
\name{normalize.max}
\alias{normalize.max}
\title{Max normalization}
\usage{
normalize.max(S)
}
\arguments{
\item{S}{matrix with the raw non normalized scores. Rows are examples and columns are classes.}
}
\value{
A score matrix with the same dimensions of \code{S}, but with scores normalized.
}
\description{
Function to normalize the scores of the flat scores matrix \code{S} per class.
}
\details{
the score of each class is normalized by dividing the score values for the maximum score of that class.
}
\examples{
data(scores);
maxnorm <- normalize.max(S);
}
