% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEMDAG.1.0.0.R
\name{constraints.matrix}
\alias{constraints.matrix}
\title{Constraints matrix}
\usage{
constraints.matrix(g)
}
\arguments{
\item{g}{a graph of class \code{graphNEL}L. Represent the hierarchy of the class}
}
\value{
a constraints matrix w.r.t the graph g
}
\description{
This function returns a matrix with two columns and as many rows as there are edges.
The entries of the first columns are the index of the node the edge cames from (i.e. children nodes), 
the entries of the second columns indicate the index of node the edge is to (i.e. parents nodes). 
Referring to a DAG this matrix defines a partial order.
}
\examples{
data(graph);
m <- constraints.matrix(g);
}
