% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEMDAG.1.0.0.R
\name{find.best.f}
\alias{find.best.f}
\title{Best hierarchical F-score}
\usage{
find.best.f(target, pred, n.round = 3, f.criterion = "F", verbose = TRUE,
  b.per.example = FALSE)
}
\arguments{
\item{target}{matrix with the target multilabels: rows correspond to examples and columns to classes.
\eqn{target[i,j]=1} if example \eqn{i} belongs to class \eqn{j}, \eqn{target[i,j]=0} otherwise}

\item{pred}{a numeric matrix with predicted values (scores): rows correspond to examples and columns to classes}

\item{n.round}{number of rounding digits to be applied to pred (\code{default=3})}

\item{f.criterion}{character. Type of F-measure to be used to select the best F-score. There are two possibilities:
\enumerate{
\item \code{F} (def.) corresponds to the harmonic mean between the average precision and recall;
\item \code{avF} corresponds to the per-example F-score averaged across all the examples.
}}

\item{verbose}{boolean. If \code{TRUE} (def.) the number of iterations are printed on stdout}

\item{b.per.example}{boolean. 
\itemize{
\item \code{TRUE}: results are returned for each example;
\item \code{FALSE}: only the average results are returned
}}
}
\value{
Two different outputs respect to the input paramenter \code{b.per.example}:
\itemize{
\item \code{b.per.example==FALSE}: a list with a single element average. A named vector with 7 elements relative to the best result in terms 
	of the F.measure: Precision (P), Recall (R), Specificity (S), F.measure (F), av.F.measure (av.F), Accuracy (A) and the best selected Threshold (T). 
	F is the F-measure computed as the harmonic mean between the average precision and recall; av.F is the F-measure computed as the average across 
examples and T is the best selected threshold;
	\item \code{b.per.example==FALSE}: a list with two elements:
		\enumerate{
			\item average: a named vector with with 7 elements relative to the best result in terms of the F.measure: Precision (P), Recall (R), 
		Specificity (S), F.measure (F), av.F.measure (av.F), Accuracy (A) and the best selected Threshold (T). 
			\item per.example: a named matrix with the Precision (P), Recall (R), Specificity (S), Accuracy (A), F-measure (F),	av.F-measure (av.F)
			and the best selected Threhold (T) for each example. Row names correspond to examples, column names correspond respectively 
		to Precision (P), Recall (R), Specificity (S), Accuracy (A), F-measure (F), av.F-measure (av.F) and the best selected Threhold (T).
		}
}
}
\description{
Function to select the best hierarchical F-score by choosing an appropriate threshold in the scores
}
\details{
All the examples having no positive annotations are discarded. The predicted scores matrix (\code{pred}) is rounded 
according to parameter \code{n.round} and all the values of \code{pred} are divided by \code{max(pred)}.
Then all the thresholds corresponding to all the different values included in \code{pred} are attempted, and the threshold 
leading to the maximum F-measure is selected.
}
\examples{
data(labels);
data(scores);
root <- root.node(g);
L <- L[,-which(colnames(L)==root)];
S <- S[,-which(colnames(S)==root)];
FMM <- find.best.f(L,S,n.round=3, f.criterion ="F", verbose=TRUE, b.per.example=TRUE);
}
