% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.fun.R
\name{normalize.max}
\alias{normalize.max}
\title{Max normalization}
\usage{
normalize.max(S)
}
\arguments{
\item{S}{a scores matrix. Rows are examples and columns are classes.}
}
\value{
A scores matrix with the scores normalized.
}
\description{
Normalize the scores of a scores matrix by dividing the score values of each class for the maximum score of the class.
}
\examples{
data(scores);
maxnorm <- normalize.max(S);
}
