% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.fun.R
\name{normalize.max}
\alias{normalize.max}
\title{Max normalization}
\usage{
normalize.max(S)
}
\arguments{
\item{S}{a score matrix. Rows are examples and columns are classes.}
}
\value{
A score matrix with the scores normalized.
}
\description{
Normalize the scores of a score matrix by dividing the score values of each class for the maximum score of the class.
}
\examples{
data(scores);
maxnorm <- normalize.max(S);
}
