% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEMDAG.1.0.0.R
\name{root.node}
\alias{root.node}
\title{Root node}
\usage{
root.node(g)
}
\arguments{
\item{g}{a graph of class \code{graphNEL}. It represents the hierarchy of the classes.}
}
\value{
name of the root node
}
\description{
Find the root node of a directed graph
}
\examples{
data(graph);
root <- root.node(g);
}
