% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.utility.R
\name{specific.annotation.list}
\alias{specific.annotation.list}
\title{Specific annotations list}
\usage{
specific.annotation.list(ann)
}
\arguments{
\item{ann}{an annotation matrix (0/1). Rows are examples and columns are the most specific functional terms. It must be a named matrix.}
}
\value{
A named list, where names of each component correspond to examples (genes) and elements of each component are the associated functional terms.
}
\description{
Build the annotation list starting from the matrix of the most specific annotations.
}
\examples{
data(labels);
spec.list <- specific.annotation.list(L);
}
